This file can be ran from a terminal.
I have updated many issues that arose from syntax changes to libraries and I have moved everything to a single python file to remove the need for jupyter notebook to be installed. I have change the filepath to work with Windows instead of Linux and removed the font pointer (line 69 of original).